Source/Scripts/devops_python/requirements.txt
mysql-connector as mariadb